Organize into a ConnectionManager with Transport layers
